+++ 08.ls_object/lib/models/ls_file.rb
@@ -1,0 +1,67 @@
+# frozen_string_literal: true
+
+require 'etc'
+require 'pathname'
+
+module Models
+  class LsFile
+    def self.all(params)
+      pattern = File.join(params.target_directory, '*')
+      flags = params.dot_match? ? [File::FNM_DOTMATCH] : []
+      paths = Dir.glob(pattern, *flags)
+      paths << File.join(params.target_directory, '..') if params.dot_match?
+      sorted_paths = params.reverse? ? paths.sort.reverse : paths.sort
+      sorted_paths.map { |path| LsFile.new(path) }
+    end
+
+    def initialize(path)
+      @pathname = Pathname.new(path)
+    end
+
+    def name
+      @pathname.basename.to_s
+    end
+
+    def block_size
+      file_stat.blocks
+    end
+
+    def file_type
+      file_stat.ftype
+    end
+
+    def permission
+      file_stat.mode.to_s(8)[-3..]
+    end
+
+    def link_count
+      @pathname.directory? ? @pathname.entries.size : 1
+    end
+
+    def owner_name
+      Etc.getpwuid(file_stat.uid).name
+    end
+
+    def group_name
+      Etc.getgrgid(file_stat.gid).name
+    end
+
+    def bytesize
+      @pathname.size
+    end
+
+    def mtime
+      @pathname.mtime
+    end
+
+    def to_s
+      name
+    end
+
+    private
+
+    def file_stat
+      @pathname.stat
+    end
+  end
+end
