+++ 08.ls_object/test/views/short_format_test.rb
@@ -1,5 +1,5 @@
 require 'minitest/autorun'
-require_relative '../../lib/ls_file'
+require_relative '../../lib/models/ls_file'
 require_relative '../../lib/views/short_format'
 
 module Views
@@ -10,9 +10,9 @@ module Views
         b.txt  d.txt
       TEXT
       ls_files = ('b'..'e').map do |name|
-        LsFile.new("#{name}.txt")
+        Models::LsFile.new("#{name}.txt")
       end
-      ls_files = [LsFile.new('aa.txt'), *ls_files]
+      ls_files = [Models::LsFile.new('aa.txt'), *ls_files]
       assert_equal expected, ShortFormat.new(ls_files).format
     end
   end
